+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.36.0] - 2018-12-17
+### Added
+- CXV1-15959 - New function to update skills proficiency of users.
+
+### Changed
+- Duplicate skills members of user to preserve proficiency values.
+
 ## [8.35.0] - 2018-12-11
 ### Added
 - function to update users effective capacity rule